api/v1/site_product/serializer.py
from rest_framework import serializers

from sitee.models import Product


class ProductSerializer(serializers.ModelSerializer):
    class Meta:
        model = Product
        fields = '__all__'
